Add strong params for editions
A lot of edition tests used the controller_attributes_for_instance method
to build a hash from the attributes of an edition which they send to the
controller method. This means a lot of attributes that we don't actually
want to be updateable are included, which raises an error because of the
setting of config.action_controller.action_on_unpermitted_parameters.

This commit stops that method being used in tests and removes it.