Catch errors raised by edition service observers
This is a quick fix to address in particular the issue that scheduled
publishing could fail if any of the observers fail. We had an problem
where the email service failed and the AuthorNotifier failed, causing
scheduled publishing to fail.

This needs to be revisited to provide a proper fix, in particular:

* the main service object should run in a transaction
* the failure of one observer should not prevent other observers from
  firing

See https://www.pivotaltracker.com/story/show/64036694