Exclude fake source URLs from mappings CSV
Some document sources have fake old URLs which were added on import.
These URLs have paths containing 'fabricatedurl' or 'placeholderunique'.
Because these URLs never existed, they should not be included as old URLs
in the mappings CSV consumed by Transition.

It looks as though the existing URLs of this type are on document
sources, but check attachment source URLs as well for safety.