Fix consultation published dates
A previous refactoring[1] mistakenly broke the behaviour of the document
history for consultations.

Consultations display the opening_at date rather than their first
publication date. To do this they rely on the Edition#first_public_at
method to provide an alternative behviour. Unfortunately the refactoring
broke this behaviour.

Added a unit test and fixed the behaviour.

[1]
https://github.com/alphagov/whitehall/commit/4981e263ac05b22af5efa9eb9eaf1ad5fa46346e