Publish editions in a transaction
we want changes to the edition to be atomic. The changes which occur in
observers are not core to the atomic operation of publishing an edition.

Furthermore, failure of an observer should not cause failure of the
publishing operation (see f67c5bdce83a), and therfore should not be part
of the transaction.