Ensure all attachment records have a type set
Currently a number of attachments in the database have a blank STI
"type" attribute. This is because we were previously creating
attachments using a nested attributes-based UI, which doesn't allow
the type to be set. We no longer use nested attributes for managing
attachments so we are free to fix this data now.