Make Attachment base class abstract
We are no longer creating attachments without the STI "type" attribute
set in the app, which allows us to reorganise attachment code such
that the file-related code is in FileAttachment and HtmlAttachment
contains code related to html attachments. The base class will now
raise an exception if an attempt is made to save an instance without
a specific type set.

Note that the attachment_data association is still on the base class,
otherwise the file-related scopes would not work on associations,
e.g. edition.attachments.with_filename(filename).