Replace questionable nested attributes massaging
This is to get rid of a deprecation warning related to the setting of
a non-existent "_delete" attribute on a model so that the negative
checkbox (uncheck to delete) has (and maintains) the correct state.

I've replaced nested attributes with a custom attribute writer for the
nation inapplicabilities that takes care of creating, updating and
deleting inapplicabilities as appropriate. A better solution would be
to use a form object, but the Edition model is so entangled that it
would require a serious amount of reworking to introduce one into the
edition workflow.