Translation links generate based on current action
This commit removes a kind of complicated helper which generated a link
to a translated version of a resource.  Given there aren't any translation
links away from the given resource page, it would be easier to generate
the links based on the current controller action.

This fixes https://www.pivotaltracker.com/story/show/62057862