Support exporting data for the detailed index.
This adds support to the rummager_export script for exporting data for
the detailed index.  Add a note on how to do this in the README.  Change
the README to suggest dumping the data for loading into the index to a
temporary file rather than piping the output of the export straight to
the bulk load script, because the export script tends to use a lot of
memory and is liable to fall over if other processes in the VM are
trying to do stuff.