Fixed email logic that infers create or update
When a first major edition is published we
enqueue a job to alert subscibers about it.
An editor updated a freshly published edition
while this job was in queue. So when this job
was picked up for processing, it added a line
to indicate that an update happened.

This is misleading as the editor marked the
update as a minor one and inspite of that
this email was sent out.

Updated the logic to use a better way to
determine if the email is for a create or
update, which even takes care of the email
job getting processed very late.