Handle changes to first_published_at in history
Editors have the ability to change the first_published_at timestamp
in subsequent editions. This means we cannot base the displayed time
on the value in the first edition. Instead we use the value in the
most recently pubished edition as the source of truth.

This also changes the query used to get all published editions as we
cannot rely on the in_reverse_chronological_order scope as that
returns editions based on the public timestamps. Document history is
concerned with the order editions were created in and not the public
timestamps, which can change based on changes to first_published_at.