Few tid-bits of info about imported documents
while working on this feature we realised that
if an editor starts working through the list
using the speed tagging interface and leaves
it midway there's no way to see how far they
went ?

this tiny change will at least provide one
place where the importer can see how many
of the documents from this import were
moved to Draft and an ordered list will
give a sense of the import order.

it also helps test the story.