Setting misbehaving citizens right
took me a while to realize that the connection
my test transaction was running within was not
the connection Import uses. when running with
other tests, my test picked-up one of these
previously opened connections which we've
removed in this commit.

because ActiveRecord transactions are per-connection,
Import was totally oblivious of the data setup in my
test causing unexpected failures.