Drop parallel test databases before running create
If the database isn't dropped in full then tables dropped in a migration
will still exist in the parallel databases.

Because the primary database was being dropped before creation this lead
to intermittent false positives where a test would fail if run against
the primary but would pass if run against one of the parallels where the
table still exists.