New class for cleaning document filter parameters
We have been getting random requests where filter parameters are being
specified as incorrect types. For example:

  {
    "action" => "index",
    "controller" => "publications",
    "departments" => "[\"department-of-energy-climate-change\"]",
    "locale" => "en",
    "page" => {
      "&#36;acunetix" => "1"
    },
    "publication_type" => "transparency-data"
  }

This new class returns a clean set of parameters that can be used by
the various search backends to filter documents.

It incorporates the correcting of array parameters that become
malformed when passed through the Facebook referrer from here:
https://github.com/alphagov/whitehall/blob/858d4d789cbdd0e9bd45e0cc47f70c0e54f7d331/app/controllers/application_controller.rb#L75