Cleanup unused code.
All the behaviour of these methods is now encapsulated in
Whitehall::DocumentFilter::CleanedParams and in
`ConsultationController`s case, calling `clean_search_filter_params`
is redundant anyway. Note that there will be a slight change in
behaviour: previously, any scaler params that were incorrectly passed
in as an array (e.g. page: []) are now dropped instead of taking the
first value of the array as the actual value. We don't generate urls
like so it doesn't seem reasonable that we should guess as to what
they should be. As long as we don't blow up when they are passed in,
all should be good.