Fix alternate style top tasks partial inclusion
The header partial is used by multiple views, not all of which want the
top tasks block and not all of which are in the same directory as the
partial.

Update the render call to use the full partial path and require the list
of tasks to be passed in instead of assuming they should be rendered.