Using @rinogo's fork of the text highlighter JS
it contains a fix for scrollbar issues.

the plugin makes an assumption that scrollbar
width is fixed at 18. whereas, this fork uses
a calculated scrollbar width.

https://github.com/rinogo/jQuery-highlightTextarea

this problem surfaces when using a trackpad and
there are no scrollbars shown. however, there is
no styling that tells the plugin that there are
no scrollbars shown, so it uses it's default 18px
padding.