Move search indexing from delayed job to sidekiq
As well as creating a new sidekiq-compatible workers to perform search
indexing and deleting of whitehall models, this also introduces a new
interface that hides the background-jobiness of the actions, i.e.
the Whitehall::SearchIndex.add and delete methods. This reduces the
number of integration points for search index actions and hides the
search index mechanisms away from the client code.

Note that the existing delayed job worker code is left in place. This
is so that any existing jobs that may be queued can still be run. The
code can be removed once the backlog of jobs has been worked off.