Cleanup delayed_job testing helpers
These are no longer required in the tests as all jobs now use sidekiq.
Sidekiq jobs are worked off inline during the test unit tests.

This also bumps the sidekiq gem so that we can take advantage of its
new test helpers to control whether jobs are performed inline or not
for specific tests and cucumber scenarios.