Add missing opt_value to ga event tracking
The 4th attribute in track event is opt_value which has to be
numeric, the 5th is the opt_noninteraction. When adding the
opt_noninteraction the opt_value was skipped so the wrong variable type
was being sent. ga ignores incorrect calls so the tracking in its
current state doesn't do anything.