Rewrite tests for clarity and honesty
The main thing I wanted to clarify is the special case for speeches.
Previously, the test was describing the mechanics without actually
explaining what the intent was. It was also making the code hard to
change because it was stubbing a protected method that actually no
longer does the right thing (we no longer set the public_timestamp on
speeches based on their delivered_on date).

Without rewriting the tests like this, the next commit would not have
been possible.