Remove redundant special case
In the past, the public_timestamp for speeches was set based on their
delivered_on timestamp, so we had to have this special case code to
ensure that speeches published the day after they were delivered still
got notifications. As of https://github.com/alphagov/whitehall/pull/552
however, we no longer do that and speeches get the same
public_timestamp as other edition classes, making this code both
dangerously confusing and redundant.