Only skip railties when compiling asssets
So it turns out active_record itself was always being loaded, because
delayed_job_activerecord was requiring it. Now that that's gone, we
need to explicitly require it here, otherwise the isbn_validation gem
falls over. This also gets rid of the reference to
DISABLE_ACTIVE_RECORD, as that doesn't appear to be in use anymore.