Add a less fragile workaround for CIP URL helpers.
The previous version relied on `edition` being available and linking to
an organisation, which wasn't always the case when creating a new page.
We now explicitly rely on a helper being available from the controller,
which is still unpleasant, but should fail more predictably and be
easier to track down.

(I still don't like this; alternative suggestions welcome.)