Only show stats announcements tab where appropriate
This feature is going through some alpha testing by the ONS so we only
restrict visibility of the tab in the main menu to a subset of users.
Long term, this menu item is unlikely to feature so prominently, but
whilst they are testing the feature, it makes sense to keep it as
accessible as possible for them.