Move info page title blanking to model
CorporateInformationPages derive their title from the page type, so the
title attribute can largely be ignored. But because edition titles are
translatable, we need to set CIP titles to a blank string so that they
are still returned by the with_translations scope (which excludes
anything with a NULL title).