Remove external link tracking
This was added to get analytics on traffic from specialist guides (now
detailed guidance) to organisation homepages when the organisations
linked in the metadata of those pages were all external. As those links
are all internal this is no longer needed.

Also this way of tracking external links is really very evil.