Fix redirector export after CIP changes.
Now that CIPs are editions, they can be exported in the same way as other
edition types, including state, http status, etc. The logic for calculating
their external URLs is already encapsulated in public_document_routes_helper,
so we can call that (for all editions) rather than trying to replicate it in the
exporter code. We do explicitly pass in the slug though since we want to use
the Unpublishing slug (again, for all editions) if that exists.