Hide the global header bar
Hide this element that is being introduced in static. Once static has
been deployed this can be removed and the custom header bar in this
project can be removed.

Hiding the new element means that deploys don't need to be syncronised
and the new colour bar can be adopted when this repo is ready.