Do not require body on CIP about pages
The "about us" information was extracted out of the "description" and
"about_us" attributes on the Organisation model and turned into a new
CorporateInformationPage type of AboutUs, setting the "summary" and
"body" respectively. Because only "description" was required on
Organisation, there are now a number of CIP AboutUs pages that do not
have a "body" and are thus invalid. This change makes an AboutUs page
valid without a body, which means editors will not be forced to supply
an AboutUs body for organisations that do not need one (e.g. closed or
except organisations)