Update the structure of metadata on document pages
Move from lots of desecrate bits of metadata to a more unified approach
with only a 'part of' or 'from' metadata.

Remove the old document_metadata method in favour of two helpers one for
part of and one for from. They both return an array of links which can
then be output in either the top or bottom of the page.

Updated the tests to ensure all the items which should be outputted for
each method is.

Moved the operational_field and location metadata items to the extra
metadata partial. They don't make sense in either part of or from and
are only used by one format.

Added classes on to each type of link which is going to be displayed so
we can continue to get analytical data for them.