Exclude worldwide orgs in org tags for now
WorldwideOrganisations are not yet registerd with panopticon so we want
to exclude them from organisation_ids, otherwise registration fails.
This has become an issue since we converted CorporateInformationPages to
an Edition type as CorporateInformationPage#organisations will return
an array containing the owning organisation, which will be either a
normal Organisation or a WorldwideOrganisation.