Change default static asset serving in dev mode
Up to now, whitehall has defaulted to serving assets from preview in
development mode. This is the opposite of how all the other apps in our
stack works. This reverses that so that whitehall expects assets to be
served from wherever Plek says they will be, but also allows the static
server to be overridden directly with an ENV var (STATIC_DEV) if
required. For convenience for those developers who are running whitehall
outside of the VM, startup.sh has been updated to set this ENV var so
assets are served from preview.