Remove lots of hardcoded URLs
This also gets rid of some special case tests that were testing the
difference between the public and admin hosts on preview and production.
As we are now solely depending on the truth as told by Plek, we don't
need to have separate tests for preview and production in most cases.