Report exceptions raised during edition workflow
The fact that we are rescue these exceptions is a long-standing smell
and a source of technical debt that we need to address. Work was started
to extract all these workflow actions into self-contained and atomic
service objects such that we would no longer need to rescue these
exceptions (see EditionPublisher, etc), but that work was not completed.

In preparation for removing these rescue_from statements, this commit
hads additional logging so that we receive notifications if/when
exceptions are raised during workflow actions. This will allow us to
better asses the likely impact of removing them now instead of waiting
until all workflow actions have had service objects extracted.