Spike: speed up news article edit form
Proof of concept of a potential technique to speed up the edit forms for
editions. At present, these take several seconds to load, making for a
pretty miserable user experience.

This commit demonstrates how using Rails caching to cache the options
used in form select fields could yield some serious speed improvements.
In almost all cases, the values in these options do not change very
often (e.g. lists of organisations, policies, etc). This caching would
have to be combined with an effective strategy for invalidating the
caches when there are changes.

In this example, the new news article form goes from loading in
approximately 5 seconds to loading in less than a 1.3 seconds.