Add ability for LinkChecker to check authed links
Because the preview environment is hidden behind HTTP basic auth, link
checking will fail for any links to other pages on the site in this
environment. This change allows us to configure domains that require
authentication such that the LinkChecker will be able to automatically
apply the auth credentials when required. This will be configured at
deploy time so that environment-specific details are kept out of the
codebase.