/government/statistics shows filtered index of statistics
Statistics are being extracted from the publications section of the site
to be displayed under /government/statistics. This commit creates a new
statistics controller.

It's more of a smoke-and-mirrors approach; The new StatisticsController
inherits from PublicationsController and gets most of it's functionality
from there.

Some thought was given into whether building a new provider in a similar
pattern to the stats announcements, but we decided that since gds hasn't
decided on what shape this kind of thing should be yet, it would be better
to take a less drastic approach for now.