Remove unoccupied role messaging
This is more often confusing than helpful. Becuase of the way we model roles
it's possible for a role to have been replaced/superceded by a new role
with a slightly different name, which technically means the old role is no
longer occupied. This is confusing for high profile roles.