Rework rake tasks for overdue editions
Tweaks the rake task for listing overdue editions and adds one to
publish those that are overdue. The main change is that we only include
those that are more than one minute overdue, which gives enough of a
buffer for the workers to complete their work before something is
considered overdue.