Add link to Services and Information page
Where an organisation displays a block of "Services and guidance" at
the top of the page, include a link through to the new Services and 
Information page for the org.

There's some fiddly CSS here, due to a style rule in another file 
overriding the padding-left value for the link. I'd like to know if this
could be implemented in a cleaner way.