Add a data migration to trim excessive slugs
We introduced a limit on slug length in ca28979, however documents
created before this time don't conform to this validation, which breaks
the crawler.

This data migration cleans up the old slugs for published editions by
updating their slugs to the new maximum length and redirecting the old
URLs to the new URLs.