Migration to update LegislativeList syntax
Versuion 3.0 of govspeak introduces a breaking change for legislative
lists, requiring a closing tag.

This migration adds a closing tag to any existing uses of
`$LegislativeList` in HTML attachment bodies in the place that the list
end would have previously been inferred.

This is done as a migration rather than a data migration to try to
ensure a minimum length of time that things look broken. It should only
be broken for the time it takes for unicorn to finish restarting.