Only serve statistics publications from /statistics
They were still routable from /publications, as well as /statistics, which is bad.
Serve them from one authorative place, and support any use of old style with redirects

Removes a couple of old tests for statistics-behaviour on publication views. The first
of which failed when addin the redirect, the 2nd 2 silently pass, as they're refuting
a selector against an empty body.