Allow removal of deleted announcements from index
Destroying statistics announcements raises an exception with real
rummager because it cannot generate the link to remove it from the
index. This is because the :statistics_announcement_dates association is
now a dependent and gets destroyed when the announcement is destroyed,
but the current announcement date is delegated to when generating the
searchable content to get the link.