Add filters to searches on land registry pages
When a search is performed on content which is owned by the Land
Registry, scope the search results to contain only land registry
content.  This is for https://www.pivotaltracker.com/story/show/77617490

Since the Land Registry transitioned, we've seen a sharp and significant
increase in the number of search results which get no result clicks.
Also, analytics shows that lots of fairly generic searches are being
performed from land-registry pages (eg, for terms such as "portal" and
"forms").  Our hypothesis is that these are searches from people who
have been redirected from the old site, and that these searches have a
high expectation that searches will be across Land Registry content.

We want to deploy a change now to filter the results accordingly.  The
results page will open with the "Organisations" filter box expanded, and
the "land-registry" organisation selected (and at the top of the list of
organisations in the filter box).  We'll then track the analytics to see
if this scope works well in this case, with a view to scoping or biasing
the search results from more organisation pages in future.