Stop incorrect date initialisation at month end
Bug manifested as a failing test on Aug 29, the tests use February as
an example. The JavaScript was creating a date, using today, then going
through and setting each of the properties, eg year, month etc. On
August 29, setting the month to February would try and create a date
“February 29” (non-leap year) which is not a valid date, so JavaScript
switches to March 1st.

* Avoid bug by creating variables from each date input then creating a
date all at once