Unscope SpecialistSector relation to Edition
Each SpecialistSector belongs to an edition. However, if the edition is 
deleted, requesting the relation returns nil as the default scope for 
the Edition model is to exclude deleted editions.

By overriding the relation to be unscoped, we can still perform 
operations on Editions related to SpecialistSectors regardless of their 
state.

This is based on prior art found in 
https://github.com/alphagov/whitehall/blob/ec83c2ea/app/models/classification_membership.rb#L25