Script for cleaning up duplicate announcements
Some code that will make it easier to delete and redirect duplicate
statistics announcements in future. It registers a redirect from a
duplicated announcement to an authoritative one and destroys the
duplicate. The script can be run from the command line:

  ./script/dedupe_stats_announcement dupe_slug authoritative_slug

The script also accepts a "-n" option to call the command as NOOP.