Rake task for generating realistic benchmarks
Adds a `test:alt_benchmarks` rake task to more accurately measure the
performance of some poorly performing pages.

This leans on the existing performance testing facilities in Rails, but
instead of running against the "test" environment, it defines a new one
called "benchmark" that is set up to behave more like production. The
main differences between "test" and "benchmark" are:

  * caching will be performed in "benchmark"
  * the database will not be dropped/rebuilt on every run

The main aim here is to more accurately model the behaviour in
production, right down to the volume of data in the database. By
preloading a dump of production data into the benchmark database, these
performance tests will give us a more accurate measure of page load
times for those pages that contain lots of form elements referencing
data from the database (e.g. select lists for organisations, related
policies, etc).

A set of reference tests have been added as an example.