Script to move content from one topic to another
This is part of https://www.pivotaltracker.com/story/show/76471962

All editions tagged with the source topic are retagged replacing the
source topic with the destination topic.

I decided to apply this to all editions including old ones since this
script is expected to be used when doing bulk re-assignment of topics;
we'll be likely to delete the source topic using the existing
specialist_sector_cleanup task after running this, so any old editions
where we didn't re-assign the topic would just lose their topic.