Add note of explanation for future-developer
The way that associated lead organisations and supporting organisations
are managed is far from ideal; instead of explicit and simple
associations for each, there is a single "organisations" association,
and various hacky methods are used to massage lead and supporting orgs
into this single association whilst maintaining an order (See the
Edition::Organisations mixin and the
:delete_absent_edition_organisations before_filter on
Admin::EditionsController for a sense of the gynmastics being
performed).

Because of this, we can't rely on the conventional rails way of
accessing unsaved lead and supporting organisations in the forms, we
must instead go through the unsaved join models. These helpers
encapsulate that lookup.